switch to natively reading .opus files
..instead of that weird .dsa file type. We can just shell out to some
standard tools like ffmpeg to convert from any audio type to an .ogg
file with opus encoding. Then it's only a matter of reading the ogg file
and extracting the opus packets from there.

The OggReader and OpusReader is still pretty fragile and wacky and
assumes a small set of config options.

Signed-off-by: Daniel Chen <danielchen97dc@gmail.com>